Revert "fix window height & rename Wrapper for input elements"
This reverts commit 7377f7a5a191b991ba36d7f5ce01df1e24f6c674.
